Ubuntu 24.04, fail2ban default is version 1.0.x , which is lack of python3 module asynchat, and this module is removed from python 3.12. Install `fail2ban 1.1 version instead`

+++ functions/F_02_PKG_05_fail2ban_01_install.sh
@@ -1,4 +1,27 @@
-apt install -y fail2ban
+local if_fail2ban_installed="$(dpkg -l fail2ban 2>/dev/null | grep -E "^ii")"
+if [[ -n "${if_fail2ban_installed}" ]]; then
+  apt purge -y fail2ban
+fi
+
+
+local fail2ban_apt_version="$(apt info fail2ban 2>/dev/null| grep -i version | grep -Eo '[[:digit:]]+\.[[:digit:]]')"
+
+if [[ "${fail2ban_apt_version}" = "1.0" ]]; then
+  cd ${TMP} && mkdir fail2ban-source
+  cd ${TMP}/fail2ban-source && wget http://kr.archive.ubuntu.com/ubuntu/pool/universe/f/fail2ban/fail2ban_1.1.0-2_all.deb
+  dpkg -i fail2ban_1.1.0-2_all.deb
+
+  cd ${TMP} && SAFE_DELETE "fail2ban-source"
+else
+  apt install -y fail2ban
+fi
+
+local if_fail2ban_installed="$(dpkg -l fail2ban 2>/dev/null | grep -E "^ii")"
+if [[ -z "${if_fail2ban_installed}" ]]; then
+  echo "-------- Failed installing package 'fail2ban' --------"
+  exit 1
+fi
+
 systemctl stop fail2ban
 systemctl disable fail2ban.service
